CHG auth util file to module
