+++ api/anubis/utils/auth/http.py
@@ -1,132 +1,9 @@
-import traceback
-from datetime import datetime, timedelta
 from functools import wraps
-from typing import Union, Optional, Callable, Any
 
-import jwt
-from flask import g, request, has_request_context
-from werkzeug.local import LocalProxy
-
-from anubis.config import config
-from anubis.models import User, TAForCourse, ProfessorForCourse
+from anubis.models import TAForCourse, ProfessorForCourse
+from anubis.utils.auth.user import get_current_user
 from anubis.utils.data import is_debug
 from anubis.utils.exceptions import AuthenticationError, AssertError
-from anubis.utils.services.logger import logger
-
-
-def get_user(netid: Union[str, None]) -> Union[User, None]:
-    """
-    Load a user by username
-
-    :param netid: netid of wanted user
-    :return: User object or None
-    """
-    if netid is None:
-        return None
-
-    # Get the user from the database
-    user = User.query.filter_by(netid=netid).first()
-
-    return user
-
-
-def get_current_user() -> Union[User, None]:
-    """
-    Load current user based on the token
-
-    :return: User or None
-    """
-    if g.get("user", default=None) is not None:
-        return g.user
-
-    # Attempt to get the token from the request
-    token = get_token()
-    if token is None:
-        return None
-
-    # Try to decode the jwt
-    try:
-        decoded = jwt.decode(token, config.SECRET_KEY, algorithms=["HS256"])
-    except Exception as e:
-        logger.error('AUTH decode error\n' + traceback.format_exc())
-        return None
-
-    # Make sure there is a netid in the jwt
-    if "netid" not in decoded:
-        return None
-
-    # Get the user from the decoded jwt
-    user = get_user(decoded["netid"])
-
-    # Cache the user in the request context
-    g.user = user
-
-    return user
-
-
-def _create_get_current_user_field(field: str) -> Callable:
-    def _func() -> Optional[Any]:
-        """
-        Load current_user.id
-        :return:
-        """
-
-        # Get current user
-        user = get_current_user()
-
-        # Make sure they exist
-        if user is None:
-            return None
-
-        # Return the user.id
-        return getattr(user, field)
-
-    return _func
-
-
-def get_token() -> Union[str, None]:
-    """
-    Attempt to get the token from the request. Both the cookie, and the
-    headers will be checked.
-
-    :return:
-    """
-
-    if not has_request_context():
-        return None
-
-    return request.headers.get("token", default=None) or request.cookies.get(
-        "token", default=None
-    ) or request.args.get('token', default=None)
-
-
-def create_token(netid: str, exp_kwargs=None, **extras) -> Union[str, None]:
-    """
-    Get token for user by netid. You can provide a dictionary
-    to the exp_kwargs to set a different expire time for this token.
-    By default it is 6 hours. If you wanted to do 6 days exp_kwargs={'days': 6}
-
-    :param exp_kwargs:
-    :param netid:
-    :return: token string or None (if user not found)
-    """
-
-    # Get user
-    user: User = get_user(netid)
-
-    if exp_kwargs is None:
-        exp_kwargs = {'hours': 6}
-
-    # Verify user exists
-    if user is None:
-        return None
-
-    # Create new token
-    return jwt.encode({
-        "netid": user.netid,
-        "exp": datetime.utcnow() + timedelta(**exp_kwargs),
-        **extras,
-    }, config.SECRET_KEY)
 
 
 def require_user(unless_debug=False):
@@ -254,6 +131,3 @@ def require_superuser(unless_debug=False):
         return wrapper
 
     return decorator
-
-
-current_user: User = LocalProxy(get_current_user)
