+++ api/anubis/views/admin/autograde.py
@@ -2,7 +2,7 @@ from flask import Blueprint, request
 from sqlalchemy.sql import or_
 
 from anubis.models import Submission, Assignment, User, InCourse
-from anubis.utils.auth import require_admin
+from anubis.utils.auth.http import require_admin
 from anubis.utils.data import req_assert
 from anubis.utils.http.decorators import json_response
 from anubis.utils.http.https import success_response, get_number_arg
