+++ api/anubis/views/admin/regrade.py
@@ -6,7 +6,7 @@ from sqlalchemy import or_
 
 from anubis.models import Submission, Assignment, User
 from anubis.rpc.batch import rpc_bulk_regrade
-from anubis.utils.auth import require_admin
+from anubis.utils.auth.http import require_admin
 from anubis.utils.data import split_chunks, req_assert
 from anubis.utils.http.decorators import json_response
 from anubis.utils.http.decorators import load_from_id
