+++ api/anubis/views/admin/seed.py
@@ -1,6 +1,6 @@
 from flask import Blueprint
 
-from anubis.utils.auth import require_superuser
+from anubis.utils.auth.http import require_superuser
 from anubis.utils.data import is_debug, req_assert
 from anubis.utils.http.decorators import json_response
 from anubis.utils.http.https import success_response
