+++ api/anubis/views/admin/students.py
@@ -1,7 +1,8 @@
 from flask import Blueprint
 
 from anubis.models import db, User, Course, InCourse, Submission, Assignment
-from anubis.utils.auth import require_admin, current_user, require_superuser
+from anubis.utils.auth.http import require_admin, require_superuser
+from anubis.utils.auth.user import current_user
 from anubis.utils.data import req_assert
 from anubis.utils.http.decorators import json_response, json_endpoint
 from anubis.utils.http.https import success_response, get_number_arg
