+++ api/anubis/views/admin/visuals.py
@@ -1,7 +1,7 @@
 from flask import Blueprint
 
 from anubis.models import Assignment, User
-from anubis.utils.auth import require_admin
+from anubis.utils.auth.http import require_admin
 from anubis.utils.data import req_assert
 from anubis.utils.http.decorators import json_response
 from anubis.utils.http.https import success_response
