+++ api/anubis/views/public/assignments.py
@@ -1,6 +1,7 @@
 from flask import Blueprint, request
 
-from anubis.utils.auth import current_user, require_user
+from anubis.utils.auth.http import require_user
+from anubis.utils.auth.user import current_user
 from anubis.utils.http.decorators import json_response
 from anubis.utils.http.https import success_response
 from anubis.utils.lms.assignments import get_assignments
