+++ api/anubis/views/public/courses.py
@@ -1,7 +1,8 @@
 from flask import Blueprint
 
 from anubis.models import db, InCourse, Course
-from anubis.utils.auth import require_user, current_user
+from anubis.utils.auth.http import require_user
+from anubis.utils.auth.user import current_user
 from anubis.utils.data import req_assert
 from anubis.utils.http.decorators import json_response
 from anubis.utils.http.https import success_response, error_response
