+++ api/anubis/views/public/lectures.py
@@ -1,6 +1,7 @@
 from flask import Blueprint, request
 
-from anubis.utils.auth import require_user, current_user
+from anubis.utils.auth.http import require_user
+from anubis.utils.auth.user import current_user
 from anubis.utils.http.decorators import json_response
 from anubis.utils.http.https import success_response
 from anubis.utils.lms.lectures import get_lecture_notes
