+++ api/anubis/views/public/profile.py
@@ -3,7 +3,8 @@ import string
 from flask import Blueprint, request
 
 from anubis.models import User, db
-from anubis.utils.auth import current_user, require_user
+from anubis.utils.auth.http import require_user
+from anubis.utils.auth.user import current_user
 from anubis.utils.data import req_assert
 from anubis.utils.http.decorators import json_response
 from anubis.utils.http.https import success_response
